requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.2
cycler==0.12.1
filelock==3.19.1
fonttools==4.61.0
fsspec==2025.9.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==11.3.0
polars==1.35.2
polars-runtime-32==1.35.2
psutil==7.1.3
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.235
ultralytics-thop==2.0.18
urllib3==2.6.0
